Multiclass classification

Run a multiclass classification algorithm on a given dataset and reference class.

classification(data, class, algorithms, rfe = FALSE, ova = FALSE,
  standardize = FALSE, sizes = NULL)

Arguments

data
data frame with rows as samples, columns as features

class
true/reference class vector used for supervised learning

algorithms
character string of algorithm to use for supervised learning. See Algorithms section for possible options.

rfe
logical; if TRUE, run Recursive Feature Elimination as a feature selection method for "lda", "rf", and "svm" algorithms.

ova
logical; if TRUE, use the One-Vs-All approach for the knn algorithm.

standardize
logical; if TRUE, the training sets are standardized on features to have mean zero and unit variance. The test sets are standardized using the vectors of centers and standard deviations used in corresponding training sets.

sizes
the range of sizes of features to test RFE algorithm

Value

The model object from running the classification algorithm

Details

Some of the classification algorithms implemented use pre-defined values that specify settings and options while others need to tune hyperparameters. "multinom" and "nnet" use a maximum number of weights of 2000, in case data is high dimensional and classification is time-consuming. "nnet" also tunes the number of nodes (1-5) in the hidden layer. "pam" considers 100 thresholds when training, and uses a uniform prior. "adaboost" calls maboost::maboost() instead of adabag::boosting() for faster performance. As a result, we use the "entrop" option, which uses the KL-divergence method and mimics adaboost.

When alg = "knn", the return value is NULL because class::knn() does not output an intermediate model object. The modelling and prediction is performed in one step. However, the class attribute "knn" is still assigned to the result in order to call the respective prediction() method. An additional class "ova" is added if ova = TRUE.

Algorithms

The classification algorithms currently supported are:

Prediction Analysis for Microarrays ("pam")

Support Vector Machines ("svm")

Random Forests ("rf")

Linear Discriminant Analysis ("lda")

Shrinkage Linear Discriminant Analysis ("slda")

Shrinkage Diagonal Discriminant Analysis ("sdda")

Multinomial Logistic Regression using

Generalized Linear Model with no penalization ("mlr_glm")

GLM with LASSO penalty ("mlr_lasso")

GLM with ridge penalty ("mlr_ridge")

Neural Networks ("mlr_nnet")

Neural Networks ("nnet")

Naive Bayes ("nbayes")

Adaptive Boosting ("adaboost")

Extreme Gradient Boosting ("xgboost")

K-Nearest Neighbours ("knn")

Examples

data(hgsc)
class <- attr(hgsc, "class.true")
classification(hgsc, class, "rf")

#> 
#> Call:
#>  randomForest(x = data, y = class) 
#>                Type of random forest: classification
#>                      Number of trees: 500
#> No. of variables tried at each split: 17
#> 
#>         OOB estimate of  error rate: 13.29%
#> Confusion matrix:
#>        DIF.C4 IMM.C2 MES.C1 PRO.C5 class.error
#> DIF.C4    113      6      3     13  0.16296296
#> IMM.C2     10     89      8      0  0.16822430
#> MES.C1      3      3    102      1  0.06422018
#> PRO.C5     13      2      3    120  0.13043478