splendid: SuPervised Learning ENsemble for Diagnostic IDentification

Provides a bootstrapping and ensemble framework for supervised learning analyses using multiclass classification algorithms for modelling, prediction, and evaluation. Predicted classes are evaluated under metrics such as log loss, AUC, F1-score, Matthew's correlation coefficient, and accuracy. Discrimination and reliability plots visualize the classifier performances. The .632+ estimator is implemented for the log loss function.

See also

Useful links:

https://github.com/AlineTalhouk/splendid

Report bugs at https://github.com/AlineTalhouk/splendid/issues