Machine learning classification algorithms

Runs a classification algorithm on a given dataset and reference class.

classification(data, class, algs)

Arguments

data
data object with rows as samples, columns as features

class
true/reference class vector used for supervised learning

algs
character string of classification algorithm to use. See Details in splendid for a list of choices.

Value

The model object from running the classification algorithm "alg"

Details

Some of the classification algorithms implemented use pre-defined values that specify hyperparameters, settings, options, etc. "multinom" and "nnet" increase the maximum number of weights used to 2000, in case data is high dimensional and classification is time-consuming. "nnet" uses 3 nodes in its hidden layer, a choice that hopefully promotes sufficient complexity in many datasets. "pamr" considers 100 thresholds when training, and uses a uniform prior. "adaboost" actually calls maboost instead of boosting because of faster performance. As a result, we use the "entrop" option, which uses the KL-divergence method and mimics adaboost.

When alg = "knn", the result is NULL because the knn does not have output an intermediate model object. The modelling and prediction is done in one step. However, a class attribute is still assigned to the result in order to enact the corresponding method in prediction.

Note

"qda" gives errors when using the hgsc dataset because there are too many variables The algorithm requires the size of every class to be greater than the number of features. A feature selection framework and certain assertion checks need to be built for this algorithm to work.

Examples

data(hgsc)
class <- stringr::str_split_fixed(rownames(hgsc), "_", n = 2)[, 2]
classification(hgsc, class, "rf")

#> 
#> Call:
#>  randomForest(x = data, y = class) 
#>                Type of random forest: classification
#>                      Number of trees: 500
#> No. of variables tried at each split: 17
#> 
#>         OOB estimate of  error rate: 13.29%
#> Confusion matrix:
#>        DIF.C4 IMM.C2 MES.C1 PRO.C5 class.error
#> DIF.C4    113      6      3     13  0.16296296
#> IMM.C2     10     89      8      0  0.16822430
#> MES.C1      3      3    102      1  0.06422018
#> PRO.C5     13      2      3    120  0.13043478