Evaluation of prediction performance

Evaluation of prediction performance on the OOB set is done using various measure for classification problems.

evaluation(x, y, plot = FALSE)

Arguments

x
actual class labels

y
predicted class labels

plot
logical; if TRUE a discrimination plot and reliability plot are shown for each class

Value

A list with one element per evaluation measure except for the cs element, which returns a list of class-specific evaluation measures.

Details

The currently supported evaluation measures include discriminatory measures like log loss, AUC, and PDI, macro-averaged PPV (Precision)/Sensitivity (Recall)/F1-score, accuracy (same as micro-averaged PPV Sensitivity/F1-score), Matthew's Correlation Coefficient (and its micro-averaged analog), and class-specific PPV/Sensitivity/F1-score/MCC.

Examples

data(hgsc)
class <- factor(attr(hgsc, "class.true"))
set.seed(1)
training.id <- sample(seq_along(class), replace = TRUE)
test.id <- which(!seq_along(class) %in% training.id)
mod <- classification(hgsc[training.id, ], class[training.id], "xgboost")
pred <- prediction(mod, hgsc, class, test.id)
evaluation(class[test.id], pred)

#> $logloss
#> [1] 0.9650857
#> 
#> $auc
#> [1] 0.912119
#> 
#> $pdi
#> [1] 0.8001893
#> 
#> $accuracy
#> [1] 0.7089947
#> 
#> $macro_ppv
#> [1] 0.7067206
#> 
#> $macro_sensitivity
#> [1] 0.7132884
#> 
#> $macro_f1
#> [1] 0.7065146
#> 
#> $mcc
#> [1] 0.6136055
#> 
#> $micro_mcc
#> [1] 0.6119929
#> 
#> $cs
#>         ppv.DIF.C4         ppv.IMM.C2         ppv.MES.C1         ppv.PRO.C5 
#>          0.7142857          0.5555556          0.7843137          0.7727273 
#> sensitivity.DIF.C4 sensitivity.IMM.C2 sensitivity.MES.C1 sensitivity.PRO.C5 
#>          0.6140351          0.6756757          0.8695652          0.6938776 
#>          f1.DIF.C4          f1.IMM.C2          f1.MES.C1          f1.PRO.C5 
#>          0.6603774          0.6097561          0.8247423          0.7311828 
#>         mcc.DIF.C4         mcc.IMM.C2         mcc.MES.C1         mcc.PRO.C5 
#>          0.5319909          0.5068871          0.7663011          0.6454394 
#>