Evaluation of prediction performance

Evaluation of prediction performance on the OOB set is done using various measure for classification problems.

evaluation(x, y)

Arguments

x
actual class labels

y
predicted class labels

Value

A list with one element per evaluation measure

Details

The currently supported evaluation measures include overall accuracy, average accuracy across all One-Vs-All confusion matrices, and macro-averaged precision, recall, and F1-score.

References

https://github.com/saidbleik/Evaluation/blob/master/eval.R

Examples

data(hgsc)
class <- stringr::str_split_fixed(rownames(hgsc), "_", n = 2)[, 2]
set.seed(1)
training.id <- sample(seq_along(class), replace = TRUE)
test.id <- which(!seq_along(class) %in% training.id)
mod <- classification(hgsc[training.id, ], class[training.id], "lda")
pred <- prediction(mod, hgsc, test.id)
evaluation(class[test.id], pred)

#> $accuracy
#> [1] 0.3544974
#> 
#> $avg_accuracy
#> [1] 0.6772487
#> 
#> $avg_precision
#> [1] 0.3780929
#> 
#> $avg_recall
#> [1] 0.3675173
#> 
#> $avg_F1
#> [1] 0.3474771
#>