One-Vs-All training approach

One-Vs-All training approach

ova_classification(data, class, algorithms, rfe = FALSE, ova = FALSE,
  standardize = FALSE, sampling = c("none", "up", "down", "smote"),
  seed_samp = NULL, trees = 100, tune = FALSE, seed_alg = NULL)

Arguments

data
data frame with rows as samples, columns as features

class
true/reference class vector used for supervised learning

algorithms
character string of algorithm to use for supervised learning. See Algorithms section for possible options.

rfe
logical; if TRUE, run Recursive Feature Elimination as a feature selection method for "lda", "rf", and "svm" algorithms.

ova
logical; if TRUE, use the One-Vs-All approach for the knn algorithm.

standardize
logical; if TRUE, the training sets are standardized on features to have mean zero and unit variance. The test sets are standardized using the vectors of centers and standard deviations used in corresponding training sets.

sampling
the default is "none", in which no subsampling is performed. Other options include "up" (Up-sampling the minority class), "down" (Down-sampling the majority class), and "smote" (synthetic points for the minority class and down-sampling the majority class). Subsampling is only applicable to the training set.

seed_samp
random seed used for reproducibility in subsampling training sets for model generation

trees
number of trees to use in "rf" or boosting iterations (trees) in "adaboost"

tune
logical; if TRUE, algorithms with hyperparameters are tuned

seed_alg
random seed used for reproducibility when running algorithms with an intrinsic random element (random forests)

Value

list of binary classifier fits on each class