One-Vs-All training approach

One-Vs-All training approach

ova_classification(data, class, algs, rfe, ova)

Arguments

data
data object with rows as samples, columns as features

class
true/reference class vector used for supervised learning

algs
character string of classification algorithm to use. See Details in splendid for a list of choices.

rfe
logical; if TRUE, run Recursive Feature Elimination as a feature selection method for "lda", "rf", and "svm" algorithms.

ova
logical; if TRUE, use the One-Vs-All approach for the knn algorithm.

Value

list of binary classifier fits on each class