Train, predict, and evaluate classification models

Train, predict, and evaluate classification models

splendid_model(data, class, n, seed = 1, algorithms = NULL, rfe = FALSE,
  ova = FALSE, threshold = 0.5, ...)

Arguments

data
data object with rows as samples, columns as features

class
true/reference class vector used for supervised learning

n
number of bootstrap replicates to generate

seed
random seed used for reproducibility in bootstrapping results

algorithms
character vector of algorithm names to use for supervised learning. See Details for possible options. This argument is NULL by default, in which case uses all implemented algorithms.

rfe
logical; if TRUE, run Recursive Feature Elimination as a feature selection method for "lda", "rf", and "svm" algorithms.

ova
logical; if TRUE, a One-Vs-All classification approach is performed for every algorithm in algorithms. The relevant results are prefixed with the string ova_.

threshold
a numeric indicating the lowest maximum class probability below which a sample will be unclassified.

...
additional arguments to splendid_model

Examples

data(hgsc)
class <- attr(hgsc, "class.true")
sl_result <- splendid_model(hgsc, class, n = 1, algorithms = "xgboost")